+++ batchDemo/src/main/java/com/sudhakar/batchDemo/configuration/BatchConfiguration.java
@@ -1,0 +1,50 @@
+package com.sudhakar.batchDemo.configuration;
+
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.StepContribution;
+import org.springframework.batch.core.job.builder.JobBuilder;
+import org.springframework.batch.core.repository.JobRepository;
+import org.springframework.batch.core.scope.context.ChunkContext;
+import org.springframework.batch.core.step.builder.StepBuilder;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.batch.repeat.RepeatStatus;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.transaction.PlatformTransactionManager;
+
+@Configuration
+public class BatchConfiguration {
+    @Bean
+    public Step step1(JobRepository jobRepository, PlatformTransactionManager transactionManager) {
+        return new StepBuilder("Step1", jobRepository)
+                .tasklet(new Tasklet() {
+                    @Override
+                    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+                        System.out.println("Hi First Step1");
+                        return RepeatStatus.FINISHED;
+                    }
+                }, transactionManager)
+                .build();
+    }
+    @Bean
+    public Step step2(JobRepository jobRepository, PlatformTransactionManager transactionManager) {
+        return new StepBuilder("Step2", jobRepository)
+                .tasklet(new Tasklet() {
+                    @Override
+                    public RepeatStatus execute(StepContribution contribution, ChunkContext chunkContext) throws Exception {
+                        System.out.println("Hi First Step2");
+                        return RepeatStatus.FINISHED;
+                    }
+                }, transactionManager)
+                .build();
+    }
+    @Bean
+    public Job firstJob(JobRepository jobRepository, Step step1, Step step2) {
+        return new JobBuilder("job1", jobRepository)
+                .start(step1)
+                .next(step2)
+                .build();
+    }
+
+}
