+++ batchDemo/src/main/resources/application.properties
@@ -1,1 +1,8 @@
 spring.application.name=batchDemo
+spring.datasource.url=jdbc:mysql://localhost:3306/batch_schema
+spring.datasource.username=springstudent
+spring.datasource.password=springstudent
+spring.datasource.platform=mysql
+spring.batch.jdbc.initialize-schema=always
+spring.batch.job.enabled=false
+server.port=8085
